+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 
 - Symbol rotation [#294](https://github.com/mapsforge/vtm/issues/294)
 - Osmagray theme [#300](https://github.com/mapsforge/vtm/issues/300)
+- OkHttp external cache [#135](https://github.com/mapsforge/vtm/issues/135)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.8.0)